+++ test/forceMove/_validSignatures.test.ts
@@ -1,7 +1,7 @@
 import {ethers} from 'ethers';
 // @ts-ignore
 import ForceMoveArtifact from '../../build/contracts/TESTForceMove.json';
-import {setupContracts, sign} from './test-helpers';
+import {setupContracts, sign} from '../test-helpers';
 
 const provider = new ethers.providers.JsonRpcProvider(
   `http://localhost:${process.env.DEV_GANACHE_PORT}`,
